add symlinks if they don't exist

+++ dot-init/symlinks.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-ln -s $(realpath ../vimrc) ~/.vimrc
-ln -s $(realpath ../vim) ~/.vim
-ln -s $(realpath ../oh-my-zsh/zshrc) ~/.zshrc
-ln -s $(realpath ../oh-my-zsh/custom/plugins) ~/.oh-my-zsh/custom/
+[ ! -e ~/.vimrc ] && ln -s $(realpath ../vimrc) ~/.vimrc
+[ ! -e ~/.vim ]   && ln -s $(realpath ../vim) ~/.vim
+[ ! -e ~/.zshrc ] && ln -s $(realpath ../oh-my-zsh/zshrc) ~/.zshrc
+[ ! -e ~/.oh-my-zsh/custom/plugins ] && ln -s $(realpath ../oh-my-zsh/custom/plugins) ~/.oh-my-zsh/custom/
